+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [4.?.?] - ??-??-????
 ### Added
 - Select a custom range in the Summary
+- Predefined income or expense categories
 
 ## [4.4.0] - 03-26-2025
 ### Added